Added Timeout for Httpx
Hi Bro
I have addes -timeout 15 to httpx to  reduce the number of false positive results hope you add this to make it more efficient.  I have seen that httpx wait for 5 sec only which is not able to detect domains that takes a longer time to load hence httpx ignores it and 20 sec is the ideal time i have seen to detect most of the domains that doesn't respond quickly.